move platform key indexer and return better errors.
