+++ service/policy/keymanagement/platform_key_indexer_test.go
@@ -1,4 +1,4 @@
-package trust
+package keymanagement
 
 import (
 	"context"
